Fixed algorithm
Now in genMaze(), eller's algorithm should have a different value in the grid for every wall that some path cannot get to. Mine would have paths that should be traceable untraceable (as in grid values in the path were different numbers). That was fixed, but my maze still does not look very maze-like. Will continue more.