Maze generation actually done
I was getting these articles where there were no horizontal lines in random spots and I couldn't find out why. This happened twice. The first reason was because of the way I initialized the grid array: I used a for loop that made every initial number the same from the top, so without choosing that a number would connect, it would already connect (as underneath row 0 column 0 (value 0) it was already initialized as 0. The second time was me using a repeated list name that I used to change the algorithm for horizontal lists so there would be no more boxes that weren't connected to the entire set.